+++ xrit-rx.py
@@ -7,11 +7,12 @@ Frontend for CCSDS demultiplexer and image generator
 
 import ast
 from argparse import ArgumentParser
+from collections import namedtuple
 from configparser import ConfigParser
-from demuxer import Demuxer
 from os import mkdir, path
 import socket
 from time import time, sleep
+from demuxer import Demuxer
 import ccsds as CCSDS
 
 
@@ -64,9 +65,12 @@ def init():
     # Load decryption keys
     load_keys()
 
-    # Create demuxer instance
+    # Create demuxer config tuple
+    dcfg = namedtuple('dcfg', 'downlink verbose dump output images xrit blacklist keys')
     output += "/" + downlink + "/"
-    demux = Demuxer(downlink, args.v, args.dump, output, output_images, output_xrit, blacklist, keys)
+
+    # Create demuxer instance
+    demux = Demuxer(dcfg(downlink, args.v, args.dump, output, output_images, output_xrit, blacklist, keys))
 
     # Check demuxer thread is ready
     if not demux.coreReady: