+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
   - Transparent enhanced image output option
   - Added check for encrypted LRIT files in ``lrit-img.py`` and ``lrit-add.py``
   - Output file type options (Image or xRIT files)
+  - Demuxer configuration tuple
 
 ### Changed
   - Default key file name **(check when upgrading from an old version)**